Highlights temporaries when stepping into/over
This commit fixes a bug that prevented temporaries from being
highlighted when their value changed as a result of stepping into/over
the next bytecode.

The bug was caused by a double update of the temporary variable's
inspector.

The first update was triggered by updating the object in
ContextVariablesInspectorToTemporaryVariablesInspectorAdapter>>#object:.
This update set a temporary variable `changed` field to true and its
morph was highlighted as expected.

But right after that, the call to temporaryVariablesInspector context:
triggered a second update of the `changed` instance variable.
This time the variable value was not changed, and the morph was updated
to hide the highlight.

The bug was fixed by adding a temporary variable to
ContextVariablesInspectorToTemporaryVariablesInspectorAdapter to track
whether the temporary variable's inspector should be updated.